Fix `SkParse::FindScalar` is locale-dependent (#9)
Which causes SVG files with decimal values to not be properly parsed in e.g. Fleet on Linux when the locale is set to one that uses `,` as a decimal separator (e.g. `de_DE.UTF-8`).
Fix is based on the similar [fix SkSVGCanvas serialization is locale-dependent](https://github.com/jetbrains/skia/commit/785eca66d6691288d3032876804aaa36ec5dbe52).